requirements.txt
ollama==0.4.8
pyyaml==6.0.1
rich==13.7.0
prompt-toolkit==3.0.43
pygments==2.17.2
typer==0.9.0 